Moving status repository and classes to components-core
