+++ components-core/src/main/java/com/adyen/checkout/components/status/api/StatusConnection.java
@@ -6,12 +6,12 @@
  * Created by caiof on 28/8/2020.
  */
 
-package com.adyen.checkout.await.api;
+package com.adyen.checkout.components.status.api;
 
 import androidx.annotation.NonNull;
 
-import com.adyen.checkout.await.model.StatusRequest;
-import com.adyen.checkout.await.model.StatusResponse;
+import com.adyen.checkout.components.status.model.StatusRequest;
+import com.adyen.checkout.components.status.model.StatusResponse;
 import com.adyen.checkout.core.api.Connection;
 import com.adyen.checkout.core.log.LogUtil;
 import com.adyen.checkout.core.log.Logger;
